+++ packages/sdk/esbuild.config.mjs
@@ -3,22 +3,28 @@ import { wasmLoader } from 'esbuild-plugin-wasm';
 import { esmConfig, cjsConfig, getPackageJson } from '../../esbuild.config.mjs';
 
 const packageJson = getPackageJson(import.meta.url);
+const peerDependencies = Object.keys(packageJson.peerDependencies);
+const dependencies = Object.keys(packageJson.dependencies);
 
-// ESM bundle
+const mathDependencies = dependencies.filter((v) =>
+  v.startsWith('@galacticcouncil/math-')
+);
+
+// ESM bundle (Embedded wasms)
 esbuild
   .build({
     ...esmConfig,
     bundle: true,
     plugins: [wasmLoader({ mode: 'embedded' })],
-    external: Object.keys(packageJson.peerDependencies),
+    external: peerDependencies,
   })
   .catch(() => process.exit(1));
 
-// CJS bundle
+// CJS bundle (External math modules)
 esbuild
   .build({
     ...cjsConfig,
     bundle: true,
-    external: Object.keys(packageJson.peerDependencies),
+    external: peerDependencies.concat(mathDependencies),
   })
   .catch(() => process.exit(1));