+++ packages/xcm-cfg/src/configs/bifrost.ts
@@ -1,8 +1,5 @@
-import {
-  BalanceBuilder,
-  ExtrinsicBuilder,
-} from '@moonbeam-network/xcm-builder';
-import { AssetConfig, ChainConfig } from '@moonbeam-network/xcm-config';
+import { BalanceBuilder } from '@moonbeam-network/xcm-builder';
+import { AssetConfig, ChainConfig, polkadot } from '@moonbeam-network/xcm-config';
 
 import { bnc, vdot, dot } from '../assets';
 import { bifrost, hydraDX } from '../chains';
@@ -48,11 +45,32 @@ const toHydraDX: AssetConfig[] = [
     fee: {
       asset: bnc,
       balance: BalanceBuilder().substrate().system().account(),
-    }
+    },
+  }),
+];
+
+const toPolkadot: AssetConfig[] = [
+  new AssetConfig({
+    asset: dot,
+    balance: BalanceBuilder().substrate().tokens().accounts(),
+    destination: polkadot,
+    destinationFee: {
+      amount: 0.003,
+      asset: dot,
+      balance: BalanceBuilder().substrate().tokens().accounts(),
+    },
+    extrinsic: ExtrinsicBuilderV2().xTokens().transfer(),
+    fee: {
+      asset: bnc,
+      balance: BalanceBuilder().substrate().system().account(),
+    },
   }),
 ];
 
 export const bifrostConfig = new ChainConfig({
-  assets: [...toHydraDX],
+  assets: [
+    ...toHydraDX,
+    ...toPolkadot,
+  ],
   chain: bifrost,
 });